requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
celery==5.5.3
redis==5.2.0
tensorflow==2.15.0
keras==2.15.0
pillow==11.3.0
requests==2.32.5
numpy==1.26.4
python-multipart==0.0.20